pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "ai4bmr-core"
description = "AI for Biomedical Research"
authors = [
    { name = "Adriano Martinelli", email = "adriano.martinelli@chuv.ch" },
]
readme = "README.md"
# due to how PEP 440 defines version matching, prefer [incl, excl) definitions like below:
requires-python = ">=3.11, <3.12"
dependencies = ["python-dotenv", "pydantic", "pydantic-settings"]
dynamic = ["version"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.entry-points."console_scripts"]

[project.urls]
repository = "https://github.com/AI4SCR/ai4scr-data"

[project.optional-dependencies]
dev = [
    "pytest", "mypy", "black", "isort", "flake8", "pre-commit", "setuptools_scm[toml]",
#    "python-semantic-release",  # rely on setuptools_scm for versioning
]

[tool.black]
line-length = 88
skip-string-normalization = false

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
force_to_top = []

[tool.mypy]
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "pytest.*"
ignore_missing_imports = true

[tool.setuptools_scm]
version_scheme = "post-release"
write_to = "src/ai4bmr_core/_version.py"  # Example path where version info could be written

#[tool.semantic_release]
## for default values check:
## https://github.com/relekang/python-semantic-release/blob/master/semantic_release/defaults.cfg
#version_variable = "package/__init__.py:__version__"
#commit_version_number = true
##upload_to_pypi = true
#version_source = "tag_only"
#branch = "main"
## configure types which should trigger minor and patch version bumps respectively
## (note that they must be a subset of the configured allowed types):
#parser_angular_allowed_types = "build,chore,ci,docs,feat,fix,perf,style,refactor,test"
#parser_angular_minor_types = "feat"
#parser_angular_patch_types = "fix,perf"
